Partial `GossipMessage` change to support multiple object types

+++ node/src/components/network/error.rs
@@ -38,5 +38,5 @@ pub enum Error {
 
     /// Message too large.
     #[error("message of {actual_size} bytes exceeds limit of {max_size} bytes")]
-    MessageTooLarge { max_size: u32, actual_size: u64 },
+    MessageTooLarge { max_size: u32, actual_size: usize },
 }
